Fix validation of key removal (#2111)
`RemoveSession` now return boolean and in case of failure it gives an error

Additionally fixed hanging DNS test.
Also speed up tests by making timeout value be a float, e.g. smaller

Fix https://github.com/TykTechnologies/tyk/issues/2108

+++ reverse_proxy.go
@@ -302,15 +302,15 @@ type ReverseProxy struct {
 	ErrorHandler ErrorHandler
 }
 
-func defaultTransport(dialerTimeout int) *http.Transport {
-	timeout := 30
+func defaultTransport(dialerTimeout float64) *http.Transport {
+	timeout := 30.0
 	if dialerTimeout > 0 {
 		log.Debug("Setting timeout for outbound request to: ", dialerTimeout)
 		timeout = dialerTimeout
 	}
 
 	dialer := &net.Dialer{
-		Timeout:   time.Duration(timeout) * time.Second,
+		Timeout:   time.Duration(float64(timeout) * float64(time.Second)),
 		KeepAlive: 30 * time.Second,
 		DualStack: true,
 	}
@@ -389,7 +389,7 @@ func (p *ReverseProxy) ServeHTTPForCache(rw http.ResponseWriter, req *http.Reque
 	return resp
 }
 
-func (p *ReverseProxy) CheckHardTimeoutEnforced(spec *APISpec, req *http.Request) (bool, int) {
+func (p *ReverseProxy) CheckHardTimeoutEnforced(spec *APISpec, req *http.Request) (bool, float64) {
 	if !spec.EnforcedTimeoutEnabled {
 		return false, spec.GlobalConfig.ProxyDefaultTimeout
 	}
@@ -397,7 +397,7 @@ func (p *ReverseProxy) CheckHardTimeoutEnforced(spec *APISpec, req *http.Request
 	_, versionPaths, _, _ := spec.Version(req)
 	found, meta := spec.CheckSpecMatchesStatus(req, versionPaths, HardTimeout)
 	if found {
-		intMeta := meta.(*int)
+		intMeta := meta.(*float64)
 		log.Debug("HARD TIMEOUT ENFORCED: ", *intMeta)
 		return true, *intMeta
 	}
@@ -451,7 +451,7 @@ func proxyFromAPI(api *APISpec) func(*http.Request) (*url.URL, error) {
 	}
 }
 
-func httpTransport(timeOut int, rw http.ResponseWriter, req *http.Request, p *ReverseProxy) http.RoundTripper {
+func httpTransport(timeOut float64, rw http.ResponseWriter, req *http.Request, p *ReverseProxy) http.RoundTripper {
 	transport := defaultTransport(timeOut) // modifies a newly created transport
 	transport.TLSClientConfig = &tls.Config{}
 	transport.Proxy = proxyFromAPI(p.TykAPISpec)
